integration tests and offer controller unit tests
