+++ application/src/main/java/pl/luczak/michal/joboffersapp/JobOffersApplication.java
@@ -3,9 +3,6 @@ package pl.luczak.michal.joboffersapp;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-import org.springframework.scheduling.annotation.EnableScheduling;
 import pl.luczak.michal.joboffersapp.resttemplate.RestTemplateConfigProperties;
 
 @SpringBootApplication
@@ -14,8 +11,6 @@ import pl.luczak.michal.joboffersapp.resttemplate.RestTemplateConfigProperties;
         RestTemplateConfigProperties.class,
         JWTConfigurationProperties.class
 })
-@EnableMongoRepositories
-@EnableJpaRepositories
 public class JobOffersApplication {
 
     public static void main(String[] args) {
