+++ application/src/test/java/pl/luczak/michal/joboffersapp/validation/controller/api/APIValidationErrorHandlerUnitTestConfig.java
@@ -1,0 +1,18 @@
+package pl.luczak.michal.joboffersapp.validation.controller.api;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+abstract class APIValidationErrorHandlerUnitTestConfig {
+
+    @Bean
+    APIValidationErrorHandler apiValidationErrorHandler() {
+        return new APIValidationErrorHandler();
+    }
+
+    @Bean
+    TestController testController() {
+        return new TestController();
+    }
+}
