+++ application/src/test/java/pl/luczak/michal/joboffersapp/validation/controller/api/TestRequest.java
@@ -1,0 +1,9 @@
+package pl.luczak.michal.joboffersapp.validation.controller.api;
+
+import jakarta.validation.constraints.NotNull;
+
+record TestRequest(
+        @NotNull(message = "testMessage for %s field")
+        Integer testValue
+) {
+}
