+++ domain/src/main/java/pl/luczak/michal/joboffersapp/offer/dto/OfferDTO.java
@@ -11,9 +11,4 @@ public record OfferDTO(
         String companyName,
         String jobName,
         String salary
-) {
-
-    public boolean isNew() {
-        return uniqueID == null;
-    }
-}
+) {}
