+++ infrastructure/adapters/offer-rest-controller/src/main/java/pl/luczak/michal/joboffersapp/OfferRestController.java
@@ -1,7 +1,6 @@
 package pl.luczak.michal.joboffersapp;
 
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
