--- infrastructure/adapters/offer-rest-controller/src/main/java/pl/luczak/michal/joboffersapp/error/OfferAlreadyExistsResponse.java
@@ -1,9 +1,0 @@
-package pl.luczak.michal.joboffersapp.error;
-
-import org.springframework.http.HttpStatus;
-
-record OfferAlreadyExistsResponse(
-        String message,
-        HttpStatus status
-) {
-}
