+++ infrastructure/adapters/offer-rest-controller/src/main/java/pl/luczak/michal/joboffersapp/error/OfferErrorResponseDTO.java
@@ -1,0 +1,9 @@
+package pl.luczak.michal.joboffersapp.error;
+
+import org.springframework.http.HttpStatus;
+
+record OfferErrorResponseDTO(
+        String message,
+        HttpStatus status
+) {
+}
