--- infrastructure/adapters/offer-rest-controller/src/main/java/pl/luczak/michal/joboffersapp/error/OfferNotFoundResponse.java
@@ -1,9 +1,0 @@
-package pl.luczak.michal.joboffersapp.error;
-
-import org.springframework.http.HttpStatus;
-
-record OfferNotFoundResponse(
-        String message,
-        HttpStatus status
-) {
-}
