+++ infrastructure/adapters/offer-rest-controller/src/test/java/pl/luczak/michal/joboffersapp/OfferSaveRequestToOfferDTOMapperTest.java
@@ -1,0 +1,16 @@
+package pl.luczak.michal.joboffersapp;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class OfferSaveRequestToOfferDTOMapperTest {
+
+    @Test
+    void fromOfferDTO() {
+    }
+
+    @Test
+    void apply() {
+    }
+}
