+++ infrastructure/adapters/offer-rest-controller/src/test/java/pl/luczak/michal/joboffersapp/error/OfferRestControllerErrorHandlerTestConfig.java
@@ -1,0 +1,18 @@
+package pl.luczak.michal.joboffersapp.error;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+abstract class OfferRestControllerErrorHandlerTestConfig {
+
+    @Bean
+    TestController testController() {
+        return new TestController();
+    }
+
+    @Bean
+    OfferRestControllerErrorHandler offerRestControllerErrorHandler() {
+        return new OfferRestControllerErrorHandler();
+    }
+}
