+++ infrastructure/adapters/offerDAO-mongodb/src/main/java/pl/luczak/michal/joboffersapp/offer/OfferDAOMongoAdapterConfig.java
@@ -3,9 +3,11 @@ package pl.luczak.michal.joboffersapp.offer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @Configuration
 @EnableMongoAuditing
+@EnableMongoRepositories
 class OfferDAOMongoAdapterConfig {
 
     @Bean
