+++ infrastructure/security/authentication/src/main/java/pl/luczak/michal/joboffersapp/SecurityConfig.java
@@ -53,13 +53,18 @@ class SecurityConfig {
                 .requestMatchers(HttpMethod.GET, "/offers/**").permitAll()
                 .requestMatchers(HttpMethod.POST, "/token").permitAll()
                 .requestMatchers(HttpMethod.POST, "/register").permitAll()
+                .requestMatchers("/v2/api-docs").permitAll()
+                .requestMatchers("/v3/api-docs/**").permitAll()
+                .requestMatchers("/swagger-resources/**").permitAll()
+                .requestMatchers("/configuration/ui").permitAll()
+                .requestMatchers("/configuration/security").permitAll()
+                .requestMatchers("/swagger-ui/**").permitAll()
+                .requestMatchers("/webjars/**").permitAll()
+                .requestMatchers("/swagger-ui.html").permitAll()
                 .anyRequest()
                 .authenticated()
                 .and()
-                .headers().cacheControl().disable()
-                .frameOptions().disable()
-                .xssProtection().disable()
-                .contentTypeOptions().disable()
+                .headers().frameOptions().disable()
                 .and()
                 .httpBasic().disable()
                 .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
