+++ pom.xml
@@ -19,6 +19,7 @@
         <module>domain</module>
         <module>infrastructure</module>
         <module>application</module>
+        <module>integration</module>
     </modules>
 
     <properties>
@@ -95,6 +96,11 @@
                 <groupId>com.fasterxml.jackson.module</groupId>
                 <artifactId>jackson-module-parameter-names</artifactId>
                 <version>${jackson-module-parameter-names.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.springdoc</groupId>
+                <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+                <version>2.0.4</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
